requirements.txt
aiogram==3.20.0.post0
fastapi==0.115.12
python-dotenv==1.1.0
uvicorn==0.34.2
